feat: course container children view
