+++ cms/djangoapps/contentstore/rest_api/v1/views/__init__.py
@@ -3,10 +3,10 @@ Views for v1 contentstore API.
 """
 from .certificates import CourseCertificatesView
 from .course_details import CourseDetailsView
-from .course_index import CourseIndexView
+from .course_index import ContainerChildrenView, CourseIndexView
 from .course_rerun import CourseRerunView
-from .course_waffle_flags import CourseWaffleFlagsView
 from .course_team import CourseTeamView
+from .course_waffle_flags import CourseWaffleFlagsView
 from .grading import CourseGradingView
 from .group_configurations import CourseGroupConfigurationsView
 from .help_urls import HelpUrlsView
@@ -14,7 +14,7 @@ from .home import HomePageCoursesView, HomePageLibrariesView, HomePageView
 from .proctoring import ProctoredExamSettingsView, ProctoringErrorsView
 from .settings import CourseSettingsView
 from .textbooks import CourseTextbooksView
-from .vertical_block import ContainerHandlerView, VerticalContainerView
+from .vertical_block import ContainerHandlerView, vertical_container_children_redirect_view
 from .videos import (
     CourseVideosView,
     VideoDownloadView,
